# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Class Chatting API",
        "version": "0.0.1",
        "description": "Dokumen OpenAPI 3.1 untuk Class Chatting.\n\nDokumentasi ini mencakup:\n- Public API di bawah prefix `/api`\n- Private API di bawah prefix `/private-api/{api_key}`\n- Firebase webhook di bawah prefix `/api/firebase`\n\nGunakan `/docs/api` untuk dokumentasi interaktif, `/docs/api.json` untuk OpenAPI JSON runtime, dan `openapi.json` untuk artefak yang dapat dikonsumsi AI, generator SDK, atau automation tooling."
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Current application"
        }
    ],
...
    "paths": {
        "/v1/regions/provinces": {
            "get": {
                "operationId": "regionsListProvinces",
                "description": "Mengambil daftar provinsi dalam format paginasi.",
                "summary": "List provinces",
                "tags": [
                    "Regions"
                ],
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Kata kunci pencarian berdasarkan kode atau nama provinsi.",
                        "schema": {
                            "type": [
                                "string",
                                "null"
                            ],
                            "maxLength": 100
                        },
                        "example": "jakarta"
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Jumlah item per halaman. Nilai yang valid antara 1 sampai 100.",
                        "schema": {
                            "type": "integer"
                        },
                        "example": 25
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated set of `ProvinceResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ProvinceResource"
                                            }
                                        },
                                        "links": {
                                            "type": "object",
                                            "properties": {
                                                "first": {
                                                    "type": [
                                                        "string",
                                                        "null"
                                                    ]
                                                },
                                                "last": {
                                                    "type": [
                                                        "string",
                                                        "null"
                                                    ]
                                                },
                                                "prev": {
                                                    "type": [
                                                        "string",
                                                        "null"
                                                    ]
                                                },
                                                "next": {
                                                    "type": [
                                                        "string",
                                                        "null"
                                                    ]
                                                }
                                            },
                                            "required": [
                                                "first",
                                                "last",
                                                "prev",
                                                "next"
                                            ]
                                        },
                                        "meta": {
                                            "type": "object",
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "minimum": 1
                                                },
                                                "from": {
                                                    "type": [
                                                        "integer",
                                                        "null"
                                                    ],
                                                    "minimum": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "minimum": 1
                                                },
                                                "links": {
                                                    "type": "array",
                                                    "description": "Generated paginator links.",
                                                    "items": {
                                                        "type": "object",
                                                        "properties": {
                                                            "url": {
                                                                "type": [
                                                                    "string",
                                                                    "null"
                                                                ]
                                                            },
                                                            "label": {
                                                                "type": "string"
                                                            },
                                                            "active": {
                                                                "type": "boolean"
                                                            }
                                                        },
                                                        "required": [
                                                            "url",
                                                            "label",
                                                            "active"
                                                        ]
                                                    }
                                                },
                                                "path": {
                                                    "type": [
                                                        "string",
                                                        "null"
                                                    ],
                                                    "description": "Base path for paginator generated URLs."
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "description": "Number of items shown per page.",
                                                    "minimum": 0
                                                },
                                                "to": {
                                                    "type": [
                                                        "integer",
                                                        "null"
                                                    ],
                                                    "description": "Number of the last item in the slice.",
                                                    "minimum": 1
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "description": "Total number of items being paginated.",
                                                    "minimum": 0
                                                }
                                            },
                                            "required": [
                                                "current_page",
                                                "from",
                                                "last_page",
                                                "links",
                                                "path",
                                                "per_page",
                                                "to",
                                                "total"
                                            ]
                                        }
                                    },
                                    "required": [
                                        "data",
                                        "links",
                                        "meta"
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
        "/v1/regions/provinces/{province}": {
            "get": {
                "operationId": "regionsShowProvince",
                "description": "Mengambil detail satu provinsi berdasarkan kode wilayah Indonesia.",
                "summary": "Get province detail",
                "tags": [
                    "Regions"
                ],
                "parameters": [
                    {
                        "name": "province",
                        "in": "path",
                        "required": true,
                        "description": "Kode provinsi Indonesia yang dipakai route model binding.",
                        "schema": {
                            "type": "string"
                        },
                        "example": "31"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`ProvinceResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ProvinceResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    }
                }
            }
        },
        "/v1/regions/provinces/{province}/regencies": {
            "get": {
                "operationId": "regionsListRegenciesByProvince",
                "description": "Mengambil daftar kabupaten atau kota yang berada dalam satu provinsi tertentu.",
                "summary": "List regencies by province",
                "tags": [
                    "Regions"
                ],
                "parameters": [
                    {
                        "name": "province",
                        "in": "path",
                        "required": true,
                        "description": "Kode provinsi Indonesia yang menjadi induk data kabupaten atau kota.",
                        "schema": {
                            "type": "string"
                        },
                        "example": "31"
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Kata kunci pencarian berdasarkan kode atau nama kabupaten atau kota.",
                        "schema": {
                            "type": [
                                "string",
                                "null"
                            ],
                            "maxLength": 100
                        },
                        "example": "selatan"
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Jumlah item per halaman. Nilai yang valid antara 1 sampai 100.",
                        "schema": {
                            "type": "integer"
                        },
                        "example": 25
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated set of `RegencyResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/RegencyResource"
                                            }
                                        },
                                        "links": {
                                            "type": "object",
                                            "properties": {
                                                "first": {
                                                    "type": [
                                                        "string",
                                                        "null"
                                                    ]
                                                },
                                                "last": {
                                                    "type": [
                                                        "string",
                                                        "null"
                                                    ]
                                                },
                                                "prev": {
                                                    "type": [
                                                        "string",
                                                        "null"
                                                    ]
                                                },
                                                "next": {
                                                    "type": [
                                                        "string",
                                                        "null"
                                                    ]
                                                }
                                            },
                                            "required": [
                                                "first",
                                                "last",
                                                "prev",
                                                "next"
                                            ]
                                        },
                                        "meta": {
                                            "type": "object",
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "minimum": 1
                                                },
                                                "from": {
                                                    "type": [
                                                        "integer",
                                                        "null"
                                                    ],
                                                    "minimum": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "minimum": 1
                                                },
                                                "links": {
                                                    "type": "array",
                                                    "description": "Generated paginator links.",
                                                    "items": {
                                                        "type": "object",
                                                        "properties": {
                                                            "url": {
                                                                "type": [
                                                                    "string",
                                                                    "null"
                                                                ]
                                                            },
                                                            "label": {
                                                                "type": "string"
                                                            },
                                                            "active": {
                                                                "type": "boolean"
                                                            }
                                                        },
                                                        "required": [
                                                            "url",
                                                            "label",
                                                            "active"
                                                        ]
                                                    }
                                                },
                                                "path": {
                                                    "type": [
                                                        "string",
                                                        "null"
                                                    ],
                                                    "description": "Base path for paginator generated URLs."
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "description": "Number of items shown per page.",
                                                    "minimum": 0
                                                },
                                                "to": {
                                                    "type": [
                                                        "integer",
                                                        "null"
                                                    ],
                                                    "description": "Number of the last item in the slice.",
                                                    "minimum": 1
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "description": "Total number of items being paginated.",
                                                    "minimum": 0
                                                }
                                            },
                                            "required": [
                                                "current_page",
                                                "from",
                                                "last_page",
                                                "links",
                                                "path",
                                                "per_page",
                                                "to",
                                                "total"
                                            ]
                                        }
                                    },
                                    "required": [
                                        "data",
                                        "links",
                                        "meta"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
...
        "/v1/regions/regencies/{regency}": {
            "get": {
                "operationId": "regionsShowRegency",
                "description": "Mengambil detail satu kabupaten atau kota berdasarkan kode wilayah Indonesia.",
                "summary": "Get regency detail",
                "tags": [
                    "Regions"
                ],
                "parameters": [
                    {
                        "name": "regency",
                        "in": "path",
                        "required": true,
                        "description": "Kode kabupaten atau kota Indonesia yang dipakai route model binding.",
                        "schema": {
                            "type": "string"
                        },
                        "example": "3171"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`RegencyResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/RegencyResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    }
                }
            }
        },
...
        "/v1/regions/districts/{district}": {
            "get": {
                "operationId": "regionsShowDistrict",
                "description": "Mengambil detail satu kecamatan berdasarkan kode wilayah Indonesia.",
                "summary": "Get district detail",
                "tags": [
                    "Regions"
                ],
                "parameters": [
                    {
                        "name": "district",
                        "in": "path",
                        "required": true,
                        "description": "Kode kecamatan Indonesia yang dipakai route model binding.",
                        "schema": {
                            "type": "string"
                        },
                        "example": "3171010"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`DistrictResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/DistrictResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ProvinceResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "code": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "regencies_count": {
                        "type": "integer"
                    },
                    "districts_count": {
                        "type": "integer"
                    }
                },
                "required": [
                    "id",
                    "code",
                    "name"
                ],
                "title": "ProvinceResource"
            },
...
        },
        "responses": {
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview."
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "required": [
                                "message",
                                "errors"
                            ]
                        }
                    }
                }
            },
...
        }
    }
}